Correction de bugs

+++ meuporg-client/source/GraphicsUtils.inl
@@ -2,5 +2,5 @@ template<typename T>
 void centerOrigin(T& x)
 {
 	sf::FloatRect bounds = x.getLocalBounds();
-	x.setOrigin(bounds.left + bounds.width / 2.f, bounds.top + bounds.height / 2.f);
+	x.setOrigin((int)(bounds.left + bounds.width / 2.f), (int)(bounds.top + bounds.height / 2.f));
 }
